Rename to `once-per-template`
