+++ .circleci/config.yml
@@ -484,7 +484,8 @@ jobs:
           command: |
             mkdir sandbox
             cd code
-            yarn multiplex-templates --cadence ci --parallel --script "yarn sandbox --no-link --no-start --no-publish"
+            yarn once-per-template --cadence ci --parallel \
+              --script "yarn sandbox --no-link --no-start --no-publish"
       - persist_to_workspace:
           root: .
           paths:
