+++ scripts/once-per-template.ts
